+++ lib/vcs/test/test__graph.ml
@@ -26,13 +26,13 @@ let%expect_test "graph" =
     let path = Eio.Path.(Eio.Stdenv.fs env / "super-master-mind.log") in
     let contents = Eio.Path.load path in
     let lines = String.split_lines contents in
-    List.map lines ~f:(fun line -> Vcs_git_cli.Log.parse_log_line_exn ~line)
+    List.map lines ~f:(fun line -> Vcs_git_provider.Log.parse_log_line_exn ~line)
   in
   let refs =
     let path = Eio.Path.(Eio.Stdenv.fs env / "super-master-mind.refs") in
     let contents = Eio.Path.load path in
     let lines = String.split_lines contents in
-    Vcs_git_cli.Refs.parse_lines_exn ~lines
+    Vcs_git_provider.Refs.parse_lines_exn ~lines
   in
   let graph = Vcs.Graph.create () in
   print_s [%sexp { node_count = (Vcs.Graph.node_count graph : int) }];