+++ lib/vcs/test/test__num_status.ml
@@ -25,7 +25,7 @@ let%expect_test "parse_exn" =
   let path = Eio.Path.(Eio.Stdenv.fs env / "super-master-mind.num-status") in
   let contents = Eio.Path.load path in
   let lines = String.split_lines contents in
-  let num_status = Vcs_git_cli.Num_status.parse_lines_exn ~lines in
+  let num_status = Vcs_git_provider.Num_status.parse_lines_exn ~lines in
   ignore (num_status : Vcs.Num_status.t);
   [%expect {||}];
   ()