+++ lib/vcs_git_blocking/src/runtime.mli
@@ -21,6 +21,6 @@
 
 type t
 
-include Vcs_git_cli.Runtime.S with type t := t
+include Vcs_git_provider.Runtime.S with type t := t
 
 val create : unit -> t