+++ lib/vcs_git_provider/src/munged_path.ml
@@ -19,23 +19,44 @@
 (*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*******************************************************************************)
 
-let%expect_test "init" =
-  Eio_main.run
-  @@ fun env ->
-  Eio.Switch.run
-  @@ fun sw ->
-  let vcs = Vcs_git.create ~env in
-  let path = Stdlib.Filename.temp_dir ~temp_dir:(Unix.getcwd ()) "vcs" "test" in
-  let repo_root =
-    Eio.Switch.on_release sw (fun () ->
-      Eio.Path.rmtree Eio.Path.(Eio.Stdenv.fs env / path));
-    Vcs.For_test.init vcs ~path:(Absolute_path.v path) |> Or_error.ok_exn
-  in
-  require_equal
-    [%here]
-    (module Absolute_path)
-    (Absolute_path.v path)
-    (Vcs.Repo_root.to_absolute_path repo_root);
-  [%expect {||}];
-  ()
+module T = struct
+  [@@@coverage off]
+
+  type t = Vcs.Num_status.Key.t =
+    | One_file of Vcs.Path_in_repo.t
+    | Two_files of
+        { src : Vcs.Path_in_repo.t
+        ; dst : Vcs.Path_in_repo.t
+        }
+  [@@deriving equal, sexp_of]
+end
+
+include T
+
+let arrow = lazy (String.Search_pattern.create " => ")
+
+let parse_exn str =
+  try
+    match String.Search_pattern.split_on (force arrow) str with
+    | [ str ] -> One_file (Vcs.Path_in_repo.v str)
+    | [ left; right ] ->
+      (match String.rsplit2 left ~on:'{' with
+       | None ->
+         Two_files { src = Vcs.Path_in_repo.v left; dst = Vcs.Path_in_repo.v right }
+       | Some (prefix, left_of_arrow) ->
+         let right_of_arrow, suffix = String.lsplit2_exn right ~on:'}' in
+         Two_files
+           { src = Vcs.Path_in_repo.v (prefix ^ left_of_arrow ^ suffix)
+           ; dst = Vcs.Path_in_repo.v (prefix ^ right_of_arrow ^ suffix)
+           })
+    | _ :: _ :: _ -> raise_s [%sexp "Too many '=>'"] [@coverage off]
+    | [] -> assert false
+  with
+  | exn ->
+    raise_s
+      [%sexp
+        "Vcs_git_provider.Munged_path.parse_exn"
+        , "invalid path"
+        , (str : string)
+        , (exn : Exn.t)]
 ;;