+++ lib/vcs_git_provider/test/test__refs.ml
@@ -33,7 +33,7 @@ let%expect_test "parse_exn" =
   let path = Eio.Path.(Eio.Stdenv.fs env / "super-master-mind.refs") in
   let contents = Eio.Path.load path in
   let lines = String.split_lines contents in
-  let refs = Vcs_git_cli.Refs.parse_lines_exn ~lines in
+  let refs = Vcs_git_provider.Refs.parse_lines_exn ~lines in
   print_s
     [%sexp
       { tags = (Vcs.Refs.tags refs : Set.M(Vcs.Tag_name).t)
@@ -56,7 +56,7 @@ let%expect_test "parse_exn" =
 let%expect_test "parse_ref_kind_exn" =
   let test_ref_kind str =
     print_s
-      [%sexp (Vcs_git_cli.Refs.Dereferenced.parse_ref_kind_exn str : Vcs.Ref_kind.t)]
+      [%sexp (Vcs_git_provider.Refs.Dereferenced.parse_ref_kind_exn str : Vcs.Ref_kind.t)]
   in
   require_does_raise [%here] (fun () -> test_ref_kind "blah");
   [%expect {| (Invalid_argument "String.chop_prefix_exn \"blah\" \"refs/\"") |}];
@@ -84,7 +84,8 @@ let%expect_test "dereferenced" =
   let test line =
     print_s
       [%sexp
-        (Vcs_git_cli.Refs.Dereferenced.parse_exn ~line : Vcs_git_cli.Refs.Dereferenced.t)]
+        (Vcs_git_provider.Refs.Dereferenced.parse_exn ~line
+         : Vcs_git_provider.Refs.Dereferenced.t)]
   in
   require_does_raise [%here] (fun () -> test "");
   [%expect {| ("Invalid ref line" "") |}];