+++ test/expect/nonraising_unit_tests.ml
@@ -53,7 +53,7 @@ let%expect_test "num stat without lines" =
   @@ fun env ->
   Eio.Switch.run
   @@ fun sw ->
-  let vcs = Vcs_git.create ~env in
+  let vcs = Vcs_git_eio.create ~env in
   let mock_revs = Vcs.Mock_revs.create () in
   let repo_root = Vcs_test_helpers.init_temp_repo ~env ~sw ~vcs in
   let () =